+++ desktop/qml/PlantingsPage.qml
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 André Hoarau <ah@ouvaton.org>
+ * Copyright (C) 2018-2019 André Hoarau <ah@ouvaton.org>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -88,11 +88,6 @@ Page {
     title: qsTr("Plantings")
     padding: 0
     Material.background: "white"
-
-    Settings {
-        id: settings
-//        property alias tableModel: plantingsView.tableHeaderModel
-    }
 
     Shortcut {
         sequences: ["Ctrl+N"]