+++ desktop/qml/PlantingsView.qml
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2018-2019 André Hoarau <ah@ouvaton.org>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; version 3.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 import QtQuick 2.11
 import QtQuick.Controls 2.4
 import QtQuick.Layouts 1.3
@@ -279,7 +295,7 @@ ListView {
         color: Material.primary
         radius: 2
     }
-    cacheBuffer: Units.rowHeight*2
+    cacheBuffer: Units.rowHeight*4
 
     implicitWidth: contentWidth
     implicitHeight: contentHeight
@@ -354,7 +370,7 @@ ListView {
                     anchors.verticalCenter: headerRow.verticalCenter
                     tristate: true
                     checkState: rowCount && checks == rowCount ? Qt.Checked
-                                                                   : (checks > 0 ? Qt.PartiallyChecked : Qt.Unchecked)
+                                                               : (checks > 0 ? Qt.PartiallyChecked : Qt.Unchecked)
                     nextCheckState: function () {
                         if (!rowCount)
                             return;
@@ -568,7 +584,6 @@ ListView {
                     rank: model.planting_rank
                     selectionMode: checks > 0
                     anchors.verticalCenter: row.verticalCenter
-                    //                                width: 24
                     width: listView.checkBoxWidth
                     round: true
                     color: model.crop_color
@@ -630,7 +645,6 @@ ListView {
                         visible: tableHeaderModel[index+2].visible
                         anchors.verticalCenter: parent.verticalCenter
                         horizontalAlignment: tableHeaderModel[index+2].alignment
-                        elide: Text.ElideRight
                         width: tableHeaderModel[index+2].width
                     }
                 }